requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
cloudinary==1.32.0
gunicorn==20.1.0